requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
cssselect==1.0.3
cycler==0.10.0
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-crontab==0.7.1
django-rest-auth==0.9.3
djangorestframework==3.9.2
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
inflect==2.1.0
jieba3k==0.35.1
kiwisolver==1.0.1
lxml==4.3.2
matplotlib==3.0.3
more-itertools==6.0.0
newspaper3k==0.2.8
nltk==3.4
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.2
Pillow==5.4.1
pluggy==0.9.0
py==1.8.0
pyparsing==2.3.1
pytest==4.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-file==1.4.3
requests-oauthlib==1.2.0
scikit-learn==0.20.3
scipy==1.2.1
simplesentiment==1.0.6
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
soupsieve==1.8
tinysegmenter==0.3
tldextract==2.2.1
twarc==1.6.3
urllib3==1.24.1
wordcloud==1.5.0
